Update all refresh scripts to work with our new Docker dev envs
Note that there is still a good amount to cleanup here, but generally,
these scripts are what I used to populate the various S3 buckets with
database dumps and files for the local dev envs to pull from. It splits
out some of the existing overall "refresh staging" scripts into more
modular onces that can be run by themselves.