Script to dump the staging Portal DB in a compressed format.
I wanted to see hte size difference and cutover the existing scripts
to dump the DB compressed and then use pg_restore in order to create the dev
DB. However, the size was 500 vs 600 MB. It's a good deal smaller, but it
wasn't enough to prioritize this project at the moment. However, we'll likely
come back to it so may as well check this in.