First pass of splitting up staging refresh into two phases, backup and restore.
We are aiming to have nightly snapshots automatically taken so they are sitting
there ready for a dev or staging refresh to be run off them.

This first pass implements the split for the Join server and creates the directory
structure we'll implement everything under.

NOTE: this disables staging refreshes for the whole env until we cutover all the apps.

TESTING:
- I tested the Join backup (snapshot) part by running backup_scripts/all_create_snapshots.sh
- I didn't want to actually refresh staging, so I ran everything in the
  refresh_staging_scripts/join_refresh.sh except the actual refresh.