Cutover Kits staging refresh to happen in two phases, nightly backups and on-demand refreshes.
With this committ, all_create_snapshots.sh script can be run in a
nightly cronjob to snapshot the Kits production databases, convert them to
staging databases, and store them in S3. It also syncs the wp-content from
production to S3. Then on-demand, the all_refresh.sh script
can be run to restore Kits staging from the latest DB backups and wp-content.

TESTING:
- Commented out the Join snapshot functionality from all_create_snapshots.sh
  leaving only the Kits related stuff and ran it. Saw that the DB snapshots
  and wp-content were updated in the relevant S3 buckets.
- Can't test the actual staging refresh b/c we have to coordinate with people
  using staging. Though I tested the commands in my local dev env.